safari-534.53-branch/LayoutTests/media/audio-constructor-src-expected.txt

Test that Audio("url") constructor loads the specified resource.

EXPECTED (audio instanceof HTMLAudioElement == 'true') OK
EVENT(loadstart)
EXPECTED (relativeURL(audio.currentSrc) == 'content/test.wav') OK
EVENT(canplaythrough)
END OF TEST
